Fix sensing_output initialization to comply with observation_space definition

+++ ai_controller/gym_collab/gym_collab/envs/ai_collab_env.py
@@ -602,11 +602,11 @@ class AICollabEnv(gym.Env):
             "item_output": {
                 "item_weight": 0,
                 "item_danger_level": 0,
-                "item_location": np.array([0, 0], dtype=np.int16)},
+                "item_location": np.array([0, 0], dtype=np.float32)},
             "messages": "",
             "neighbors_output": {
                 "neighbor_type": 0,
-                "neighbor_location": np.array([0, 0], dtype=np.int16)},
+                "neighbor_location": np.array([0, 0], dtype=np.float32)},
             "objects_held": 0,
             "strength": strength
         }
